requirements/requirements.txt
Django==3.1.7
psycopg2==2.8.6
whitenoise==5.2.0
webp==0.1.0
Pillow==8.1.2
python-decouple==3.4